+++ src/main/java/ru/yandex/practicum/filmorate/model/Review.java
@@ -11,11 +11,11 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Builder(toBuilder = true)
 public class Review {
-    private Integer reviewId;
+    private Long reviewId;
     @NotNull
     private String content;
     private Boolean isPositive;
-    private Integer userId;
-    private Integer filmId;
+    private Long userId;
+    private Long filmId;
     private Integer useful = 0;
 }
